+++ src/PaymongoServiceProvider.php
@@ -3,6 +3,11 @@
 namespace Luigel\Paymongo;
 
 use Illuminate\Support\ServiceProvider;
+use Luigel\Paymongo\Commands\WebhookAddCommand;
+use Luigel\Paymongo\Commands\WebhookListCommand;
+use Luigel\Paymongo\Commands\WebhookToggleCommand;
+use Luigel\Paymongo\Middlewares\PaymongoValidateSignature;
+use Luigel\Paymongo\Signer\Signer;
 
 class PaymongoServiceProvider extends ServiceProvider
 {
@@ -11,36 +16,10 @@ class PaymongoServiceProvider extends ServiceProvider
      */
     public function boot()
     {
-        /*
-         * Optional methods to load your package assets
-         */
-        // $this->loadTranslationsFrom(__DIR__.'/../resources/lang', 'laravel-paymongo');
-        // $this->loadViewsFrom(__DIR__.'/../resources/views', 'laravel-paymongo');
-        // $this->loadMigrationsFrom(__DIR__.'/../database/migrations');
-        // $this->loadRoutesFrom(__DIR__.'/routes.php');
-
         if ($this->app->runningInConsole()) {
             $this->publishes([
                 __DIR__.'/../config/config.php' => config_path('paymongo.php'),
             ], 'config');
-
-            // Publishing the views.
-            /*$this->publishes([
-                __DIR__.'/../resources/views' => resource_path('views/vendor/laravel-paymongo'),
-            ], 'views');*/
-
-            // Publishing assets.
-            /*$this->publishes([
-                __DIR__.'/../resources/assets' => public_path('vendor/laravel-paymongo'),
-            ], 'assets');*/
-
-            // Publishing the translation files.
-            /*$this->publishes([
-                __DIR__.'/../resources/lang' => resource_path('lang/vendor/laravel-paymongo'),
-            ], 'lang');*/
-
-            // Registering package commands.
-            // $this->commands([]);
         }
     }
 
@@ -49,12 +28,22 @@ class PaymongoServiceProvider extends ServiceProvider
      */
     public function register()
     {
-        // Automatically apply the package configuration
+        if ($this->app->runningInConsole()) {
+            $this->commands([
+                WebhookListCommand::class,
+                WebhookAddCommand::class,
+                WebhookToggleCommand::class,
+            ]);
+        }
+
         $this->mergeConfigFrom(__DIR__.'/../config/config.php', 'paymongo');
 
-        // Register the main class to use with the facade
         $this->app->singleton('paymongo', function () {
             return new Paymongo;
         });
+
+        $this->app->bind(Signer::class, config('paymongo.signer'));
+
+        $this->app['router']->aliasMiddleware('paymongo.signature', PaymongoValidateSignature::class);
     }
 }