+++ tests/SourceTest.php
@@ -25,17 +25,17 @@ class SourceTest extends BaseTestCase
 
         $this->assertInstanceOf(Source::class, $source);
 
-        $this->assertEquals('source', $source->getType());
+        $this->assertEquals('gcash', $source->source_type);
 
-        $this->assertEquals('gcash', $source->getSourceType());
+        $this->assertEquals('source', $source->type);
 
-        $this->assertEquals(100.00, $source->getAmount());
+        $this->assertEquals(100.00, $source->amount);
 
         $this->assertEquals([
             'success' => 'http://localhost/success',
             'failed' => 'http://localhost/failed',
-            'checkout_url' => $source->getCheckoutUrlRedirect(),
-        ], $source->getRedirect());
+            'checkout_url' => $source->redirect['checkout_url'],
+        ], $source->redirect);
     }
 
     /** @test */
@@ -56,16 +56,15 @@ class SourceTest extends BaseTestCase
 
         $this->assertInstanceOf(Source::class, $source);
 
-        $this->assertEquals('source', $source->getType());
+        $this->assertEquals('grab_pay', $source->source_type);
+        $this->assertEquals('source', $source->type);
 
-        $this->assertEquals('grab_pay', $source->getSourceType());
-
-        $this->assertEquals(100.00, $source->getAmount());
+        $this->assertEquals(100.00, $source->amount);
 
         $this->assertEquals([
             'success' => 'http://localhost/success',
             'failed' => 'http://localhost/failed',
-            'checkout_url' => $source->getCheckoutUrlRedirect(),
-        ], $source->getRedirect());
+            'checkout_url' => $source->redirect['checkout_url'],
+        ], $source->redirect);
     }
 }