+++ CHANGELOG.md
@@ -2,6 +2,25 @@
 
 All notable changes to `laravel-paymongo` will be documented in this file
 
-## 1.0.0 - 201X-XX-XX
+## 1.3.0 (2020-10-31)
 
-- initial release
+### Added
+
+-   Added `getData()` for all the models.
+-   It can now access the properties of the data directly. (eg. `$token->id`)
+-   Added magic methods to all the models to get the specific data. Using the `get` prefix for the method.
+
+#### Example:
+
+```php
+// Get the ID of the token.
+$id = $token->getId();
+
+// Get the billing details of the payment method.
+$paymentMethod->getBilling();
+
+// Get the billing name of the payment method.
+$paymentMethod->getBillingName();
+```
+
+-   Added artisan commands for webhooks.